created new grid layout, refactored contact and profile pages
